third_party/WebKit/LayoutTests/fast/canvas/alpha-expected.txt

Series of tests for canvas alpha

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx1.getContextAttributes().alpha is true
PASS imgData1.data[0] is 0
PASS imgData1.data[1] is 0
PASS imgData1.data[2] is 0
PASS imgData1.data[3] is 0
PASS ctx2.getContextAttributes().alpha is true
PASS imgData2.data[0] is 0
PASS imgData2.data[1] is 0
PASS imgData2.data[2] is 0
PASS imgData2.data[3] is 0
PASS ctx3.getContextAttributes().alpha is false
PASS ctx4.getContextAttributes().alpha is true
PASS imgData4.data[0] is 0
PASS imgData4.data[1] is 0
PASS imgData4.data[2] is 0
PASS imgData4.data[3] is 0
PASS attrs.alpha is true
PASS ctx4.getContextAttributes().alpha is true
PASS imgData4.data[0] is 0
PASS imgData4.data[1] is 0
PASS imgData4.data[2] is 0
PASS imgData4.data[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
   
